Initial atomic ring code
Seems to work out fine, tests show about 12.38M items/sec with an atomic
ring on an Intel E6660, which is about one item every 194 cycles, which
seems okay.  By this point I've tested the atomic rings with about a few
billion items and haven't had an error yet, except for when I hand
modified the assembly to make sure they would happen and be caught
correctly if the memory consistency was relaxed.  (They were, and the
test programs both created and caught the errors as designed.)

On the other hand, x86_64 has surprisingly strong memory guarantees, so
it'll be more comfortable when similar tests are run on platforms with
weaker guarantees and see it still work.

Meanwhile, I'm going to start using this code now.