Update docs and file header
As always, after pushing code, there's a bunch of small formatting and
wording changes that were bothering me in the documentation.  This
commit updates the README.md file to include these changes.

Also the file header for aring.c was the same as the file header for
aring.h because surprise surprise after I wrote the header file, I just
copied it to create the .c, deleted the macro guards and then filled in
the blanks with code.  I completely forgot to re-write the file header,
so this commit fixes that too.

This commit does not change a single line of code.  Only comments and
docs.